move check scripts to check dir
