+++ typescript/infra/scripts/check/check-validator-announce.ts
@@ -1,10 +1,9 @@
 import { ChainMap, defaultMultisigConfigs } from '@hyperlane-xyz/sdk';
 import { eqAddress } from '@hyperlane-xyz/utils';
 
-import { isEthereumProtocolChain } from '../src/utils/utils.js';
-
-import { getArgs, withChains } from './agent-utils.js';
-import { getEnvironmentConfig, getHyperlaneCore } from './core-utils.js';
+import { isEthereumProtocolChain } from '../../src/utils/utils.js';
+import { getArgs, withChains } from '../agent-utils.js';
+import { getEnvironmentConfig, getHyperlaneCore } from '../core-utils.js';
 
 const minimumValidatorCount = 3;
